refactor(gabinet/calendar): import AppointmentIndicator from canonical module (#1441)
Swap the inline re-imports of `AppointmentIndicator` from `./appointment-card`
to its canonical source in `./appointment-indicators` across the four calendar
views. The re-export in `appointment-card.tsx` is preserved as-is.
