+++ src/components/gabinet/calendar/calendar-day-by-employee-view.tsx
@@ -6,7 +6,7 @@ import { DroppableSlot } from "./droppable-slot";
 import { useDragToCreate } from "./use-drag-to-create";
 import { useCurrentTime } from "@/hooks/use-current-time";
 import { cn } from "@/lib/utils";
-import type { AppointmentIndicator } from "./appointment-card";
+import type { AppointmentIndicator } from "./appointment-indicators";
 
 interface Appointment {
   _id: string;
